backend/requirements.txt
fastapi
uvicorn
google-generativeai
PyMuPDF
python-multipart
python-dotenv
